0.1.0
Created correct file path suited for this project, added and employee parent class that takes the github, phone numb, this needs to be updated in the future however I wanted to get the basic structure of the app down, Next created an engineer class that extends the properties from the employee class, finally inside of the index.js the inquirer correctly runs threw all prompts for each selection, and ensures the used has no more employees to add before stopping, all of the instantiations of the class objects are being pushed to an objects array then rendered with a function that executes a forEach method. all parameters are passed in functions and data is correctly extrapolated with best practices such as object destructuring. Base structure correctly functions and generates HMTL as intended.