0.2.3
added in css to make card display as a card to the screen, updated index.js functions renderData would now pass the full object for each element to the render html function that now destructures it inside of the function so that template litterals can be used to assign all values.  Updated the Render Html function to have better indentation in the html it contains, as well as the div to have the class of card all for styles.css.