0.3.0
updated lib to have intern file that contains intern class and a manager file that contains manager class, added in dist folder that contains the css, and html, files, updated style.css to corectly center a tags as well. updated all lib files to have a type, this was done for the update on index.js, all types (company position) are selected by app function passed to decided function a conditional checks the type then passes it to the correct ql with type param, from there after the question line has run create object function will be called with data and the type variable we have been continuously passing, create object destructures the object param then a conditional checks if the new object to be created should be engineer, manager, or intern, from the type we have been passing. and then pushes that to an objects array then calls more emp if no is selected render data function is called render data takes the objectarr with the foreach method and uses fs.appendFile to fill each portion of the index.html, using the renderhtml function. which defines the html with template literals. when complete object destructuring was used on the forEach variable, to console log that the html card was sucessfully generated with the type of position that was created.