0.4.0
added src folder that contains helper.js broke down long code file and seperated out the file write to helper.js and imported it to index.js. changed the way render data functions by adding an html variable where every time the foreach method is called on object array it will += the card class with template literals. this was a fix to prevent the declaration of the html and header, body, or just full html file from rendering repatatively and now inserts the card data into an html strucutre then it is written.