0.5.0
Updated all classes to use the correct properties and get methods, and to properly call constructors and super if they are a extended from the employee parent class. updated index.js question lines for each class, so that they ask all questions require for their relative class, fixed create object,  so that the correct values were destructured and instantiations of new objects call the correct class with the correct arguments. updated the html variable inside render data to use a turnary opperator to decided if it will render at anything. if it does it calls the relative method on the element variable from the foreach callback.